+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## Version 2.0
 
 * Renamed TypeDef -> Schema
+* Removed deprecated settings